+++ src/main/java/by/students/grsu/controller/LotController.java
@@ -17,6 +17,7 @@ import org.springframework.web.servlet.ModelAndView;
 
 import javax.annotation.PostConstruct;
 import javax.servlet.ServletRequest;
+import javax.servlet.http.HttpServletRequest;
 import java.util.List;
 
 @Controller
@@ -53,17 +54,21 @@ public class LotController {
 //    }
 
     @RequestMapping(value = "/{a_id}/addLot", method = RequestMethod.GET)
-    public ModelAndView addLot(@PathVariable("a_id") Integer a_id, SecurityContextHolderAwareRequestWrapper contextHolder) {
-        ModelAndView mv = new ModelAndView("addLot");
-        mv.addObject("a_id", a_id);
-        try {
-            List<Item> items = itemService.getFreeItemsByOwner(contextHolder);
-            mv.addObject("items", items);
-        }catch (Exception e){
-            mv.addObject("errMessage", "Internal error " + e.getMessage()+ ". Sorry.");
+    public ModelAndView addLot(@PathVariable("a_id") Integer a_id, SecurityContextHolderAwareRequestWrapper contextHolder, HttpServletRequest request) {
+        ModelAndView mv;
+        if(request.isUserInRole("ADMIN") || request.isUserInRole("SELLER")) {
+            mv = new ModelAndView("addLot");
+            mv.addObject("a_id", a_id);
+            try {
+                List<Item> items = itemService.getFreeItemsByOwner(contextHolder);
+                mv.addObject("items", items);
+            } catch (Exception e) {
+                mv.addObject("errMessage", "Internal error " + e.getMessage() + ". Sorry.");
+            }
+            mv.addObject("back", a_id + "/lotList");
+        }else{
+            mv = new ModelAndView("redirect:/" + a_id + "/lotList");
         }
-
-        mv.addObject("back", a_id + "/lotList");
         return mv;
     }
 
@@ -92,28 +97,30 @@ public class LotController {
 
     @RequestMapping(value = "/{a_id}/saveLot", method = RequestMethod.GET)
     public String lotInfo(@PathVariable("a_id") Integer a_id,
-                          ModelMap model, ServletRequest request){
-        String[] parameters = request.getParameterValues("items");
-        int[] ids = new int[parameters.length];
-        for(int i = 0; i < parameters.length; i++){
-            ids[i] = Integer.parseInt(parameters[i]);
-        }
+                          ModelMap model, HttpServletRequest request){
+        if(request.isUserInRole("ADMIN") || request.isUserInRole("SELLER")) {
+            String[] parameters = request.getParameterValues("items");
+            int[] ids = new int[parameters.length];
+            for (int i = 0; i < parameters.length; i++) {
+                ids[i] = Integer.parseInt(parameters[i]);
+            }
 
-        try {
-            LotInfo newLot = lotService.createLot(a_id, request.getParameter("name"), Double.parseDouble(request.getParameter("price")), Double.parseDouble(request.getParameter("min_price")), ids);
-            return "redirect:/" + a_id + "/" + newLot.getID() + "/lotInfo";
-        }catch(AuctionException e){
-            model.addAttribute("errMessage", "Internal error " + e.getCode() + ". Sorry.");
-            System.out.println(e.getMessage());
-            //mb return to error page
-        } catch (Exception e) {
-            model.addAttribute("errMessage", "Internal error " + e.getMessage() + ". Sorry.");
-            System.out.println(e.getMessage());
-            //mb return to error page
+            try {
+                LotInfo newLot = lotService.createLot(a_id, request.getParameter("name"), Double.parseDouble(request.getParameter("price")), Double.parseDouble(request.getParameter("min_price")), ids);
+                return "redirect:/" + a_id + "/" + newLot.getID() + "/lotInfo";
+            } catch (AuctionException e) {
+                model.addAttribute("errMessage", "Internal error " + e.getCode() + ". Sorry.");
+                System.out.println(e.getMessage() + " controller");
+                //mb return to error page
+            } catch (Exception e) {
+                model.addAttribute("errMessage", "Internal error " + e.getMessage() + ". Sorry.");
+                System.out.println(e.getMessage() + " controller");
+                //mb return to error page
+            }
         }
 
         //model.addAttribute("back", a_id + "/lotList");
-        return "error";
+        return "redirect:/" + a_id + "/lotList";
     }
 
     @RequestMapping(value = "/{a_id}/{l_id}/lotInfo", method = RequestMethod.GET)
@@ -132,43 +139,43 @@ public class LotController {
     }
 
     @RequestMapping(value = "/{a_id}/deleteLot", method = RequestMethod.GET)
-    public ModelAndView deleteLot(@PathVariable("a_id") Integer a_id) {
-        ModelAndView mv = new ModelAndView("deleteLot");
-        try {
-            AuctionInfo auc = auctionService.getAuctionWithLots(a_id);
-            //List<Lot> lots = lotService.getLotsByAuctionId(a_id);
-            mv.addObject("auction", auc);
-        }catch (Exception e){
-            mv.addObject("errMessage", "Internal error " + e.getMessage()+ ". Sorry.");
+    public ModelAndView deleteLot(@PathVariable("a_id") Integer a_id, HttpServletRequest request) {
+        ModelAndView mv;
+        if(request.isUserInRole("ADMIN") /*|| request.isUserInRole("SELLER")*/) {
+            mv = new ModelAndView("deleteLot");
+            try {
+                AuctionInfo auc = auctionService.getAuctionWithLots(a_id);
+                mv.addObject("auction", auc);
+            } catch (Exception e) {
+                mv.addObject("errMessage", "Internal error " + e.getMessage() + ". Sorry.");
+            }
+            mv.addObject("back", a_id + "/lotList");
+        }else{
+            mv = new ModelAndView("redirect:/" + a_id + "/lotList");
         }
-
-        //mv.addObject("num", new IntegerWrapper());
-        mv.addObject("back", a_id + "/lotList");
         return mv;
     }
 
     @RequestMapping(value = "/{a_id}/deleteLotPart2", method = RequestMethod.GET)
     public String itemInfo(@PathVariable("a_id") Integer a_id,
-                           ModelMap model, ServletRequest request) {
-        int num = Integer.parseInt(request.getParameter("lot"));
-
-        try {
-            Lot lot = lotService.getLotById(num);
-            //TODO check if user is admin?
+                           ModelMap model, HttpServletRequest request) {
+        if(request.isUserInRole("ADMIN") /*|| request.isUserInRole("SELLER")*/) {
+            int num = Integer.parseInt(request.getParameter("lot"));
+            try {
+                Lot lot = lotService.getLotById(num);
 //            if(!item.getOwner().equals(user.getUsername())){
 //                model.addAttribute("errMessage", "Sorry, you can't delete this item, because it's not yours.");
 //                List<Item> items = itemService.getItemsByOwner(user);
 //                model.addAttribute("items", items);
 //                return "deleteItem";
 //            }
-            lotService.deleteLot(num);
-            itemService.freeItemsByLot(num);
-        }catch (Exception e){
-            model.addAttribute("errMessage", "Internal error " + e.getMessage()+ ". Sorry.");
-            return "deleteItem";
+                lotService.deleteLot(num);
+                itemService.freeItemsByLot(num);
+            } catch (Exception e) {
+                model.addAttribute("errMessage", "Internal error " + e.getMessage() + ". Sorry.");
+                return "deleteItem";
+            }
         }
-
-        //model.addAttribute("back", "freeItems");
 
         return "redirect:/" + a_id + "/lotList";
     }
